Move HE logic to HeLogic
